# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Technical Test ",
        "description": "documentation for API Technical Test",
        "contact": {
            "email": "maszahid1@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/technical-test/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/auth/forgot-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Endpoint ini untuk mendapatkan token lupa password",
                "operationId": "a1fca6b1d9516ba1e0c1809f0e5f6b10",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "email yang terdaftar",
                                        "type": "email",
                                        "example": "ybrakus@example.net"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "ybrakus@example.net"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Akan memberikan token lupa password"
                    },
                    "201": {
                        "description": "Email tidak ditemukan"
                    },
                    "422": {
                        "description": "Some variable is required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/update-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Endpoint ini untuk mengubah password",
                "operationId": "df941cb39cf0489e89aadeed17279de4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "property untuk mengizinkan menggunakan method PUT",
                                        "type": "string",
                                        "example": "put"
                                    },
                                    "email": {
                                        "description": "email yang terdaftar",
                                        "type": "email",
                                        "example": "ybrakus@example.net"
                                    },
                                    "token": {
                                        "description": "token yang didapat ketika melakukan lupa password",
                                        "type": "string",
                                        "example": "kjabkdaisdgihiaushd"
                                    },
                                    "password": {
                                        "description": "Password baru, terdiri dari minimal 7 karakter, huruf besar dan kecil, angka dan simbol, dan bukan password yang sudah digunakan sebelumnya",
                                        "type": "password",
                                        "example": "Password3!"
                                    },
                                    "password_confirmation": {
                                        "description": "ulangi password",
                                        "type": "password",
                                        "example": "Password3!"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "_method": "PUT",
                                    "email": "ybrakus@example.net",
                                    "token": "asdasdasdadasda",
                                    "password": "Password3!",
                                    "password_confirmation": "Password3!"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Akan memberikan pesan berhasil mengubah password"
                    },
                    "201": {
                        "description": "Password gagal diubah/ password sudah pernah digunakan"
                    },
                    "422": {
                        "description": "Some variable is required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}